Don't use version 3.5 of WiX
Wix 3.5 doesn't seem to work well with Wine and it ends up combining
files that have the same size. This obviously breaks the .msi. The
README now recommends to use version 3.0 and the Makefile has a check
to verify that two files with the same size are correctly stored with
different contents in the cab.